requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.127.0
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
loguru==0.7.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pgvector==0.4.2
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.10.1
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-multipart==0.0.21
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.241
ultralytics-thop==2.0.18
urllib3==2.6.2
uvicorn==0.40.0
redis==5.0.8

